==> 2-cleaning.ipynb <==
# Notebook 2:  Data Cleaning 

Sometimes through the scraping or manual download from the website, unnecessary columns or cells with no value occur.  These should be deleted before using them in the model. 
Additonally the salaries need to be cleaned. For that we transform them into a numerical format and normalize them in respect to the salary limit of each year. (The limit is usually increased every year by the NBA.)

## Imports

%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import pandas as pd

from bs4 import BeautifulSoup, Comment
import requests
import lxml
import unicodedata
import json, pickle

## Setting up the cleaning of the players data

def strip_accents_and_punctuation(text):
    '''Normalize player name spellings'''
    try:
        text = unicode(text, 'utf-8')
    except NameError: # unicode is a default on python 3 
        pass
    text = unicodedata.normalize('NFD', text)\
           .encode('ascii', 'ignore')\
           .decode("utf-8")
    return str(text).replace('.','').replace(',','').replace("'",'')

# (1) Merge Scraped Data

Merging datasets

#loading all the different dataframes, which we want to combine into one

df2022 = pd.read_csv('./data/df2022_raw.csv');
df2023salaries = pd.read_csv('./data/dfcurrentsalaries.csv');
df2023stats = pd.read_csv('./data/dfcurrentstats.csv');
df2023urls = pd.read_csv('./data/dfplayer_to_url.csv');
df2021_merged = pd.read_csv('./data/df2021_merged.csv');
df2023_salary = pd.read_csv('./data\df2023_salary.csv'); 


## a. Update df2022 with 2023 salaries

# Merge df2022_merged with df2021_merged to get "PrevSal"
df2022 = df2022.merge(df2021_merged[['Name', 'Salary']], on='Name', how='left')

df2022.rename(columns={'Salary': 'PrevSal'}, inplace=True)
df2022.rename(columns = {'CurrentSalary':'Salary'})

# Merge the merged_df with df2023_salary to get "NextSal"
df2022 = df2022.merge(df2023_salary[['Player', '2023-24']], left_on='Name', right_on='Player', how='left')

df2022.rename(columns={'2023-24': 'NextSal'}, inplace=True)

# Adding new columns, which will be used later on, for differentiating between years.
df2022['NameYear'] = '2022'
df2022['PrevYear'] = '2021'
df2022['NextYear'] = '2023'

# Dropping redundant columns, which are generated through the merging.
df2022.drop('Player', axis=1, inplace=True)
df2022.drop('Unnamed: 0.1', axis=1, inplace=True)
df2022.drop('Unnamed: 0', axis=1, inplace=True)

df2022.head(5)

## b. Update df2022 with 2023 salaries and player urls

df2022_merged = pd.merge(df2023stats, df2023salaries, left_on="Player", right_on="Name", how="left")

df2022_merged = pd.merge(df2022, df2023urls, left_on="Name", right_on="Name", how="left")


df2022_merged = df2022_merged.rename(columns={'Player': 'Name'})

# Dropping different columns due to the merging.
df2022_merged.drop('Unnamed: 0', axis=1, inplace=True)
df2022_merged.drop('ID_x', axis=1, inplace=True)
df2022_merged.drop('ID_y', axis=1, inplace=True)
#df2022_merged.drop('Unnamed: 0_x', axis=1, inplace=True)
#df2022_merged.drop('Unnamed: 0_y', axis=1, inplace=True)

df2022_merged.head(5)

## Save merged data

df2022_merged.to_csv(r'C:\Users\Vincenzo\Documents\Master 2. Semester\GutHub_Sofi\BigDataNBA\data\df2022_merged.csv')

# (2) Clean salary data

## Load files

df2020_merged = pd.read_csv('./data/df2020_merged.csv')
df2021_merged = pd.read_csv('./data/df2021_merged.csv')
df2022 = pd.read_csv('./data/df2022_merged.csv')

# Dropping one column, which is being generated every time one loads the dataframe. (Unknown why that is the case.)
df2022.drop('Unnamed: 0', axis=1, inplace=True)

df2022.head(5)

## Delete unnecessary entries

Some data should be deleted, as they are not representative or not complete enough to receive a proper result. 

These would include players, who have not played a single game in the season and those who not received a salary for the upcoming season. This is because sometimes players receive a huge salary increasing by signing a new contract, due to the salary limitations set by the NBA. (E.g. if a player is out of his first contract with around $1-2 million ,he can negotiate a contract, which is tenfold.)
Meaning that players at the end of a contract, are much more dificult to analyse. First we would need to accurately forecast what the salary of a player will be, which we could only base on performance indicators. (Due to the scope of the project and data availability.) The same indicators we would then also use to determine if the salary is justified by the performance. 

Consequently we are discarding those players.


# We are only looking into rows, which have a value for "CurrentSalary".
df2022 = df2022[df2022['CurrentSalary'].notnull()]

#  1. Mask for players who played at least one game.

df2022['G'] = df2022['G'].apply(lambda s: pd.to_numeric(s, errors='coerce'))
mask1 = df2022['G'].isna()
df2022 = df2022[~mask1]

#  2. Mask for players who have a salary for next year.

mask2  = (df2022['NextSal'] == '0') | df2022['NextSal'].isna()  
df2022 = df2022[~mask2]
df2022.sample(5)

### Clean salries and normalize by salary cap

In order to fully normalize our data in regards to salary, we use the limitation for the salary, set by the NBA, from 2023 for all the years. This is a simpliflying approach.
We had to use a lot of transformations and with many hours of try and error, we managed to make the code work. It is definetely improvable, but we did not dare to touch it again.
Following the slogan "never touch a running system".

# Convert 'PrevYear' and 'NextYear' columns to integers
df2022['PrevYear'] = df2022['PrevYear'].astype(int)
df2022['NextYear'] = df2022['NextYear'].astype(int)

def clean_salary(sal, year):
    # Adds all salaries earned per year into one salary (in millions of dollars)
    #   "< $Minimum" is treated as $0

    if isinstance(sal, float):
        return 0
    sal = str(sal)  # Convert sal to string if it's not already    

    newlist = [0]
    for elem in sal.split('$'):
        try:
            elem = elem.replace('<','').replace(',','').replace('(TW)','').replace('Minimum','0')
            #elem =  float(elem)
            newlist.append(float(elem))
        except:
            continue
            
    salary = sum(newlist)/1000000
    
    return salary

# Salary Cap Data taken from spotrac.com/nba/cba
salarycap = {2016:  94.143, 2017: 99.093,   2018: 101.869,  2019:  109.14,   
             2020:  109.14,    2021: 112.414,   2022: 123.655,  2023: 136.021}   

def clean_and_norm_salary(sal, year):
    #Adds all salaries earned per year into one salary (in millions of dollars)
    # < $Minimum" is treated as $0
    # Then normalizes by salary cap (in 2023 salary cap dollars)

    if isinstance(sal, float):  # Check if sal is not a string
        return 0  # Return 0 for non-string values

    sal = str(sal)  # Convert sal to string if it's not already

    newlist = [0]
    for elem in sal.split('$'):
        try:
            elem = elem.replace('<','').replace(',','').replace('Minimum','0')
            elem =  float(elem)
            newlist.append(elem)
        except:
            continue
            
    salary = sum(newlist)/1000000
    salary = salary*salarycap[2023]/salarycap[year]
        
    return salary

    df2022_merged = df2022_merged.reset_index(drop=True)

# Calculate the previous and next years
df2022['PrevYear'] = df2022['PrevYear']
df2022['NextYear'] = df2022['PrevYear'] + 1

# Apply the functions to entire columns using vectorized operations
df2022['PrevSalClean'] = df2022.apply(lambda row: clean_salary(row['PrevSal'], row['PrevYear']), axis=1)
df2022['NextSalClean'] = df2022.apply(lambda row: clean_salary(row['NextSal'], row['NextYear']), axis=1)
df2022['PrevSalNorm'] = df2022.apply(lambda row: clean_and_norm_salary(row['PrevSal'], row['PrevYear']), axis=1)
df2022['NextSalNorm'] = df2022.apply(lambda row: clean_and_norm_salary(row['NextSal'], row['NextYear']), axis=1)

cols = ['NextSal', 'NextSalClean', 'NextSalNorm']
print(df2022[cols].sample(8))

for index, row in df2022.iterrows():
    
    salary = df2022.loc[index, 'CurrentSalary']
    try:
        df2022.loc[index, 'SalClean'] = clean_salary(CurrentSalary, 2023)
    except:
        df2022.loc[index, 'SalClean'] = np.nan
        
    try:
        df2022.loc[index, 'SalNorm'] = norm_salary(CurrentSalary, 2023)
    except:
        df2022.loc[index, 'SalNorm'] = np.nan

df2022.sample(3)

## Save cleaned data

df.to_csv(r'/Users/richardsihombing/Documents/BigDataNBA/data/cleaned_pastyears.csv')
df2021.to_csv(r'/Users/richardsihombing/Documents/BigDataNBA/data/cleaned_currentyear.csv')



#df.to_csv(r'C:\Users\Vincenzo\Documents\Master 2. Semester\GutHub_Sofi\BigDataNBA\data\cleaned_pastyears.csv')
df2022.to_csv(r'C:\Users\Vincenzo\Documents\Master 2. Semester\GutHub_Sofi\BigDataNBA\data\cleaned_currentyear.csv')
